02-household-queries/requirements.txt
chromadb
jq
langchain_google_genai
langchain
langchain_community
langchain_text_splitters
python-dotenv